Stop adding "commit-id" tags to commits
SPR should not modify commits automatically. Instead, introduce
patch-id, a surrogate commit-id for the local system. Commits without a
commit-id will be considered WIP and will never be pushed remotely.

To mark a commit as ready to be pushed, add a "commit-id:xxxxxxxx" tag
yourself.

Hack: panics if it would reword a commit.
Hack: instead of calculating a patch-id, it just uses the commit hash.